SplashActivity added. Swipe delete added for recycler view. RecipesResponse changes. Shared preference added.

+++ app/src/main/java/com/example/eportal/model/RecipesResponse.kt
@@ -350,7 +350,7 @@ data class UserRatings(
 	val countPositive: Int? = null,
 
 	@field:SerializedName("score")
-	val score: Int? = null,
+	val score: Double? = null,
 
 	@field:SerializedName("count_negative")
 	val countNegative: Int? = null
